[BEAM-5985] Refactor: move smoke load test suite to separate file
