+++ .test-infra/jenkins/job_LoadTests_Java.groovy
@@ -218,79 +218,3 @@ PhraseTriggeringPostCommitBuilder.postCommitJob(
     loadTestsBuilder.loadTest(delegate, testConfiguration.jobDescription, testConfiguration.runner, CommonTestProperties.SDK.JAVA, testConfiguration.jobProperties, testConfiguration.itClass.CommonTestProperties.TriggeringContext.PR)
   }
 }
-
-def smokeTestConfigurations = [
-        [
-                title        : 'GroupByKey load test Direct',
-                itClass      : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
-                runner       : CommonTestProperties.Runner.DIRECT,
-                jobProperties: [
-                        publishToBigQuery: true,
-                        bigQueryDataset  : 'load_test_SMOKE',
-                        bigQueryTable    : 'direct_gbk',
-                        sourceOptions    : '{"numRecords":100000,"splitPointFrequencyRecords":1}',
-                        stepOptions      : '{"outputRecordsPerInputRecord":1,"preservesInputKeyDistribution":true}',
-                        fanout           : 10,
-                        iterations       : 1,
-                ]
-        ],
-        [
-                title        : 'GroupByKey load test Dataflow',
-                itClass      : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
-                runner       : CommonTestProperties.Runner.DATAFLOW,
-                jobProperties: [
-                        project          : 'apache-beam-testing',
-                        tempLocation     : 'gs://temp-storage-for-perf-tests/smoketests',
-                        publishToBigQuery: true,
-                        bigQueryDataset  : 'load_test_SMOKE',
-                        bigQueryTable    : 'dataflow_gbk',
-                        sourceOptions    : '{"numRecords":100000,"splitPointFrequencyRecords":1}',
-                        stepOptions      : '{"outputRecordsPerInputRecord":1,"preservesInputKeyDistribution":true}',
-                        fanout           : 10,
-                        iterations       : 1,
-                ]
-        ],
-        [
-                title        : 'GroupByKey load test Flink',
-                itClass      : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
-                runner       : CommonTestProperties.Runner.FLINK,
-                jobProperties: [
-                        publishToBigQuery: true,
-                        bigQueryDataset  : 'load_test_SMOKE',
-                        bigQueryTable    : 'flink_gbk',
-                        sourceOptions    : '{"numRecords":100000,"splitPointFrequencyRecords":1}',
-                        stepOptions      : '{"outputRecordsPerInputRecord":1,"preservesInputKeyDistribution":true}',
-                        fanout           : 10,
-                        iterations       : 1,
-                ]
-        ],
-        [
-                title        : 'GroupByKey load test Spark',
-                itClass      : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
-                runner       : CommonTestProperties.Runner.SPARK,
-                jobProperties: [
-                        sparkMaster      : 'local[4]',
-                        publishToBigQuery: true,
-                        bigQueryDataset  : 'load_test_SMOKE',
-                        bigQueryTable    : 'spark_gbk',
-                        sourceOptions    : '{"numRecords":100000,"splitPointFrequencyRecords":1}',
-                        stepOptions      : '{"outputRecordsPerInputRecord":1,"preservesInputKeyDistribution":true}',
-                        fanout           : 10,
-                        iterations       : 1,
-                ]
-        ]
-]
-
-PhraseTriggeringPostCommitBuilder.postCommitJob(
-        'beam_Java_LoadTests_Smoke',
-        'Run Java Load Tests Smoke',
-        'Java Load Tests Smoke',
-        this
-) {
-  description("Runs load tests in \"smoke\" mode to check if everything works well")
-  commonJobProperties.setTopLevelMainJobProperties(delegate, 'master', 120)
-
-  for (testConfiguration in smokeTestConfigurations) {
-    loadTestsBuilder.loadTest(delegate, testConfiguration.title, testConfiguration.runner, CommonTestProperties.SDK.JAVA, testConfiguration.jobProperties, testConfiguration.itClass, CommonTestProperties.TriggeringContext.PR)
-  }
-}
